requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
greenlet==1.1.2
msgpack==1.0.4
neovim==0.3.1
Pillow==9.1.1
PyMySQL==1.0.2
pynvim==0.4.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
